Move Character action to separate service
